+++ tests/testthat/test_crs.R
@@ -1,14 +1,5 @@
 test_that("crs", {
 
-  # Attempt to handle GDAL3+/PROJ6+ warnings
-  has_thin_PROJ6_warnings <-
-    getNamespaceVersion("rgdal") >= as.numeric_version("1.5.8")
-
-  if (has_thin_PROJ6_warnings) {
-    prev_warnings <- rgdal::get_thin_PROJ6_warnings()
-    rgdal::set_thin_PROJ6_warnings(TRUE)
-  }
-
   #--- Input test cases
   test_crs <- data.frame(
     epsg = c(
@@ -17,7 +8,7 @@ test_that("crs", {
       # NAD27(USA) was 2163; however, it was deprecated by https://epsg.org
       # with change id "2020.020" and replaced by 9311.
       # Yet, 9311 is not yet available as of
-      # rgdal v1.5.16 / GDAL 3.1.1 / PROJ 6.3.1
+      # GDAL 3.1.1 / PROJ 6.3.1
       NAD27_USA = 2163,
       NAD83_USA = 4269,
       NAD83_2011_USA = 6318
@@ -51,11 +42,7 @@ test_that("crs", {
     expect_s3_class(sf::st_crs(epsg), expected_class)
     expect_s3_class(sf::st_crs(txt_epsg), expected_class)
 
-    tmp_spCRS <- if (rgdal::new_proj_and_gdal()) {
-      sp::CRS(SRS_string = txt_epsg)
-    } else {
-      sp::CRS(proj4_epsg)
-    }
+    tmp_spCRS <- sp::CRS(SRS_string = txt_epsg)
     expect_s3_class(sf::st_crs(tmp_spCRS), expected_class)
 
 
@@ -80,11 +67,7 @@ test_that("crs", {
     expect_identical(crs_units(epsg), expected_unit)
     expect_identical(crs_units(txt_epsg), expected_unit)
 
-    tmp_spCRS <- if (rgdal::new_proj_and_gdal()) {
-      sp::CRS(SRS_string = txt_epsg)
-    } else {
-      sp::CRS(proj4_epsg)
-    }
+    tmp_spCRS <- sp::CRS(SRS_string = txt_epsg)
     expect_identical(crs_units(tmp_spCRS), expected_unit)
 
     expect_identical(
@@ -101,11 +84,6 @@ test_that("crs", {
     )
 
     expect_identical(crs_units(r), expected_unit)
-  }
-
-  # Clean up
-  if (has_thin_PROJ6_warnings) {
-    rgdal::set_thin_PROJ6_warnings(prev_warnings)
   }
 })