+++ NEWS.md
@@ -1,10 +1,10 @@
 # rSW2st v0.2.0-9000
 
 * Github Actions are triggered for `release/**` branches in addition to `main`.
-* `r-lib` Github Actions updated to `v2`.
+* `r-lib` Github Actions updated to `v2` (#6; @dschlaep).
 * `get_data_dims()` now returns consistently an integer vector.
 * Linting updated to `lintr` >= 3 and
-  lint workflow switched from package tests to Github Action (#5).
+  lint workflow switched from package tests to Github Action (#5; @dschlaep).
 * `isoline_from_raster()` is now based on the `stars` and `sf` packages
   (removing a dependency on `rgeos`; addressing #3; @dschlaep).
 * `variogram_range()` is now based on the `stars` and `sf` packages