+++ NEWS.md
@@ -1,4 +1,4 @@
-# rSW2st v0.2.0-9000
+# rSW2st v0.2.0
 
 * Github Actions are triggered for `release/**` branches in addition to `main`.
 * `r-lib` Github Actions updated to `v2` (#6; @dschlaep).